benches: ray collision: multi threaded
Divide the vector of objects into chunks that are operated on and sent to the threadpool for parallel execution.

Single Threaded: 709.63 us
Multi Threaded (1): 6.1259 ms
Multi Threaded (2): 831.96 us
Multi Threaded (3): 194.07 us

3.5 times speed up to run it multi threaded.

When number of objects in the scene is 100000, the benchmark crashes for multithreaded/3 not sure why, must find out.