lib: trace_ray: fix: wrong light propagation
The light intensity cannot increase when the object is not
emissive. Only a fraction of the light (can be all of the light as
well) can be propagated forward. For this, each channel (rgb) must be
multiplied with light that is traced until then. This works only
because the light when scattered at a surface is between 0 and 1.