lib: shade_hit: fix: wo and wi are inverted
Outgoing ray direction must be the inverse of the current ray since
the current ray are travelling from camera into the scene and the BSDF
need not care about that. It must receive only the outgoing direction.

BSDF returns the incoming ray direction at the point of intersection
but for the next ray that is shot in the opposite direction (into the
scene), thus need to take the inverse of `wi`.