bvh: fix: ray_cast: closest intersection not used
Since ray_cast_traverse() might go in an arbitrary order for the
nodes, it is possible for the callback to be called on a element that
is actually further away after it has already got a closer point. So
the RayHitData::dist got from the callback must be lesser than the
already computed RayHitData::dist.