main: fix: viewport change leads to wrong ray casts
When the viewport size is smaller than the framebuffer size (when any
panel is open), the cursor must also the appropriate offset.

GLFW is top left as (0, 0), so need to subtract the left panel width
and top panel height from the cursor location that is with respect to
the window with and height.