main: camera shortcuts now require alt+num
This is mainly to prevent the camera from moving when typing in some
number. This should not be the fix for this but it works for now. The
correct fix would be to find a way to check if some text box in the
GUI is active and disable the necessary shortcuts. So it would involve
making the shortcuts context sensitive.