path_trace: medium: add support for medium information
This is useful for implementing glass shaders or similar shaders. This
is still not compatible with volumes since it is possible to have
refraction and volume mediums in one single region of space and for
this a list of mediums would be required and some sort of probability
distribution to choose the best medium for the next ray.