path_trace: bsdfs: lambert: store color as vec3 instead of vec4
There is no point storing transparency in a bsdf. The transparency
part (alpha channel) should be a separate parameter since the
sample (wi) will need to be chosen appropriately using this alpha
value. The color should always be thought of as intensity of red,
green and blue spectra. It doesn't make sense to have intensity of
alpha.