many: ColorPicker, UIData, Texture Access for Path Tracer
Many changes. Introduces color picker which either stores a single
color or a texture id. UI for this requires texture list thus DrawUI
now takes associated type ExtraData. For most instances, UIData is
set for DrawUI unless something specific is required.

UIData has all it's fields in a Arc<RwLock<_>> so that references and
their lifetimes don't need to be handled.

The overall design for this is not great, any additional features will
require considerable refactoring or it will sacrifice code
readability. Will refactor when possible.