path_trace: bsdfs: refraction: fix: crash when more exits > entries
In non manifold meshes, it is possible for the ray to exit a medium
more times than it enters the medium, this should lead to no sample
since there is no correct way to handle it. Better to not contribute
light for such a sample than to crash on the user.

In manifold (and non manifold) meshes, the normals can be inverted and
thus the ray would "exit" when going from air into the mesh before
anything else leading the same problem.