testing_framework: working directory support
std::process::Command::current_dir() works differently on Windows and
Unix. On Unix, the current directory is set before spawning the child
process which can lead to incorrect file path for child process but on
Windows it works as expected, sets the current working directory after
spawning the child process. A simple workaround for this is to use
absolute paths.

See https://github.com/rust-lang/rust/issues/37868 for more details.