rasterize: camera: get_raycast_direction_uv() and sensor support
Get ray cast direction given the UVs on the sensor. This requires that
the camera have some notion of a sensor.

Sensor can be thought of as a plane whose normal is along the camera's
front and is located at a distance equal to the focal length of the
camera away from the camera position.