rasterize: texture: pixels start from bottom left row wise
This is necessary change to stay consistent in the OpenGL world. There
are 2 ways to handle this. One is to have a helper function everywhere
that converts between the different coordinate systems but this is
annoying to use and it is possible to forget to use it. Having such a
helper function for OpenGL as well will involve programming a
"include"/"module" style approach for the shaders which is time
consuming and not highest priority as of right now.

The other method is to ensure that there is some consistency in the
way the pixels are stored, so anything that interfaces with OpenGL has
the pixels starting at bottom left row wise. This is not the most
elegant solution but the simplest one to implement as of right now
without hacks.