embree: fix: normals from embree are not correct
Since the normals from embree are not correct, they might be inverted,
need to interpolate the values as needed. Plus, embree would not know
the shading type of object, so might use vertex normals (so many
different ways to calculate this as well) or face normals so at some
point it is bound to go wrong. Best is to interpolate them based on
the bary coords provided.