object: rebuild_bvh_if_needed()
This is to ensure BVH is not rebuilt everytime just in case something
has changed, instead it does so only when something that actually
changes the state that requires a BVH rebuild occurs but deleting the
BVH.

It might make sense to add support for storing at what state (maybe
through some hash) the BVH was built. This can even evolve to creating
BVH in multiple states and choosing the state that is most appropriate
for that particular BVH query.