main: switch srgb rendering globally
This does mean that the previously set colours will not look proper
but it is worth it. For the most obvious ones, converting from srgb to
linear should make them look consistent, one would think that since
they were initially rendered as linear, the conversion should actually
be linear to srgb but that makes it much brighter. Need to spend time
at some point to figure out color stuff properly.