util: remove offset_of macro, use memoffset crate instead
Dereferencing a null pointer is undefined behaviour and this needs a
lot more complex way to figure out the offset. It is easier to use
memoffset crate instead.

One interesting thing to note is that, it is possible to implement
offset_of macro if the struct has a default implementation and is not
costly to create. Vulkano (the crate) seems to do this.