path_trace: take lock on samples_per_wavelength_per_pixel fewer times
The lock on samples_per_wavelength_per_pixel can be taken lesser
number of times by bundling the read lock in the case of creating the
image and write lock in the case of updating the number of samples.

This leads to a pretty good speedup, no exact numbers but it is
considerably faster.