path_trace: new parameter: store_rendered_image_every_sample
If true, will store the progressive render to rendered_image every
sample. If false, store the final render to rendered_image only when
stopping the render or all samples of the image are done.

For the diamond pentagonal prism scene (spectral rendering), it had a
speedup of 1.5x when rendering headless.